backend-render/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
# Supabase database
supabase==1.2.0
pandas==2.1.1
